# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "2SB_Kanban",
        "description": "Documentation de l'API",
        "contact": {
            "email": "clemsobonding@gmail.com"
        },
        "license": {
            "name": "ADE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subtasks": {
            "get": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Lister toutes les sous-tâches avec filtres",
                "operationId": "61fa9200c37aeb7628a117531763048e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "assigned_to",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_task_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des sous-tâches"
                    }
                }
            },
            "post": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Créer une sous-tâche",
                "operationId": "73f84ab45ad18bb9f12c490d6ef739f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "parent_task_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "integer"
                                    },
                                    "parent_task_id": {
                                        "type": "integer"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sous-tâche créée"
                    }
                }
            }
        },
        "/api/subtasks/{id}": {
            "get": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Afficher une sous-tâche",
                "operationId": "0b04b57d2c71c8f0b9990f5ead40132a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sous-tâche trouvée"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Mettre à jour une sous-tâche",
                "operationId": "ad4863a1816812b2cbd0aa774d29e615",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "integer"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sous-tâche mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Supprimer une sous-tâche",
                "operationId": "5b0d8a11f91c6d5439f05dc97ded8ce7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimée avec succès"
                    }
                }
            }
        },
        "/api/subtasks/{id}/status": {
            "put": {
                "tags": [
                    "SubTasks"
                ],
                "summary": "Changer le statut d'une sous-tâche",
                "operationId": "007ee3cc112aeefa549a826921089bf9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour"
                    }
                }
            }
        },
...
    },
...
}